feat(client): Implement virtual tile grid for map rendering
Refactored the JavaFX client's rendering engine to use a high-performance "virtual tile grid" as described by the original author.

- Replaced the inefficient method of creating new ImageView objects on every frame with a system that recycles a fixed-size grid of ImageViews for each of the three map layers.
- Implemented multiple layered panes (ground, objects, alpha layers) to ensure correct Z-ordering of map tiles and entities, allowing player sprites to render correctly between layers (e.g., under treetops).
- This change resolves the "blank screen" issue and significantly improves rendering performance and stability.

Fixes compilation error by changing the `createSprite` method signature to accept double coordinates for smooth entity positioning.

+++ DuskFiles/Clientz3/duskz/client/fx/DataManagerFX.java
@@ -209,7 +209,7 @@ public class DataManagerFX extends DataManager {
 		return new ImageView();
 	}
 
-	public ImageView createSprite(String set, int tileid, int tilex, int tiley, int tilewidth, int tileheight) {
+	public ImageView createSprite(String set, int tileid, double tilex, double tiley, int tilewidth, int tileheight) {
 		try {
 			ImageSet[] tiles = sets.get(set);
 			for (int i = 0; i < tiles.length; i++) {
